Configure Network Interface for Xenial
Add a task that will properly configure the network interface for
xenial. This is a little sloppy but should be easy to break it out
into something that is abstracted to work easily with different OS.
However the way things are setup right now several other things will
need to be changed around as well.